+++ examples/basics.js
@@ -1,14 +1,20 @@
-var fonts = {
-	Roboto: {
-		normal: 'fonts/Roboto-Regular.ttf',
-		bold: 'fonts/Roboto-Medium.ttf',
-		italics: 'fonts/Roboto-Italic.ttf',
-		bolditalics: 'fonts/Roboto-MediumItalic.ttf'
-	}
-};
+var pdfmake = require('../js/index'); // only during development, otherwise use the following line
+//var pdfmake = require('pdfmake');
 
-var pdfmake = require('../js/index');
-pdfmake.setFonts(fonts);
+var Roboto = require('../fonts/Roboto');
+pdfmake.addFonts(Roboto);
+
+// or you can define the font manually:
+/*
+pdfmake.addFonts({
+	Roboto: {
+		normal: '../fonts/Roboto/Roboto-Regular.ttf',
+		bold: '../fonts/Roboto/Roboto-Medium.ttf',
+		italics: '../fonts/Roboto/Roboto-Italic.ttf',
+		bolditalics: '../fonts/Roboto/Roboto-MediumItalic.ttf'
+	}
+});
+*/
 
 var docDefinition = {
 	content: [
@@ -20,6 +26,6 @@ var docDefinition = {
 var now = new Date();
 
 var pdf = pdfmake.createPdf(docDefinition);
-pdf.write('pdfs/basics.pdf');
-
-console.log(new Date() - now);
+pdf.write('pdfs/basics.pdf').then(() => {
+	console.log(new Date() - now);
+});