+++ examples/qrCode.js
@@ -1,18 +1,12 @@
-var fonts = {
-	Roboto: {
-		normal: 'fonts/Roboto-Regular.ttf',
-		bold: 'fonts/Roboto-Medium.ttf',
-		italics: 'fonts/Roboto-Italic.ttf',
-		bolditalics: 'fonts/Roboto-MediumItalic.ttf'
-	}
-};
+var pdfmake = require('../js/index'); // only during development, otherwise use the following line
+//var pdfmake = require('pdfmake');
 
-var pdfmake = require('../js/index');
-pdfmake.setFonts(fonts);
+var Roboto = require('../fonts/Roboto');
+pdfmake.addFonts(Roboto);
 
 var greeting = 'Can you see me';
 var url = 'http://pdfmake.org';
-var longText = 'The amount of data that can be stored in the QR code symbol depends on the datatype (mode, or input character set), version (1, …, 40, indicating the overall dimensions of the symbol), and error correction level. The maximum storage capacities occur for 40-L symbols (version 40, error correction level L):'
+var longText = 'The amount of data that can be stored in the QR code symbol depends on the datatype (mode, or input character set), version (1, …, 40, indicating the overall dimensions of the symbol), and error correction level. The maximum storage capacities occur for 40-L symbols (version 40, error correction level L):';
 
 
 function header(text) {
@@ -40,11 +34,11 @@ var docDefinition = {
 		header('same long text with fit = 100 and alignment = right'),
 		{ qr: longText, fit: 150, alignment: 'right' },
 	]
-}
+};
 
 var now = new Date();
 
 var pdf = pdfmake.createPdf(docDefinition);
-pdf.write('pdfs/qrCode.pdf');
-
-console.log(new Date() - now);
+pdf.write('pdfs/qrCode.pdf').then(() => {
+	console.log(new Date() - now);
+});