+++ webpack.config.js
@@ -21,7 +21,7 @@ module.exports = {
 	},
 	resolve: {
 		alias: {
-			fs: path.join(__dirname, './src/browser-extensions/virtual-fs.js')
+			fs: path.join(__dirname, './src/browser-extensions/virtual-fs-cjs.js')
 		}
 	},
 	node: {
@@ -46,6 +46,8 @@ module.exports = {
 									},
 									modules: false,
 									useBuiltIns: 'usage',
+									// TODO: after fix in babel remove corejs version and remove core-js dependency in package.json
+									corejs: "3.0.0",
 									loose: true
 								}
 							]
@@ -83,6 +85,8 @@ module.exports = {
 									},
 									modules: false,
 									useBuiltIns: 'usage',
+									// TODO: after fix in babel remove corejs version and remove core-js dependency in package.json
+									corejs: "3.0.0",
 									loose: true
 								}
 							]
@@ -91,7 +95,14 @@ module.exports = {
 					}
 				}
 			},
-			{ test: /pdfMake.js$/, loader: 'expose-loader?pdfMake', include: [path.join(__dirname, './src/browser-extensions')] },
+			{
+				test: /pdfMake.js$/,
+				loader: 'expose-loader',
+				options: {
+					exposes: 'pdfMake',
+				},
+				include: [path.join(__dirname, './src/browser-extensions')]
+			},
 			/* temporary bugfix for FileSaver: added hack for mobile device support, see https://github.com/bpampuch/pdfmake/issues/1664 */
 			/* waiting to merge and release PR https://github.com/eligrey/FileSaver.js/pull/533 */
 			{