+++ CHANGELOG.md
@@ -4,4 +4,7 @@
 
 - Port code base to ES6+
 - Unify interface for node and browser **(breaking change)**
+- All methods return promise instead of using callback **(breaking change)**
 - Change including virtual font storage in client-side **(breaking change)**
+- Support for loading font files via URL adresses (https:// or http:// protocol) in Node.js (client and server side now)
+- Removed gulp